Make dynamic require work
