+++ packages/admin/src/app/api/integrations/[...routes]/route.ts
@@ -1,18 +1,11 @@
 import { registerRoutes } from '@arkw/core';
 
-import { getFramework } from '@/lib/framework-utils';
+import { framework } from '@/lib/framework-utils';
 
-let handler;
-
-async function init() {
-  const framework = await getFramework();
-  if (!framework) {
-    throw new Error('Framework is undefined');
-  }
-
-  const handler = registerRoutes({ framework });
+if (!framework) {
+  throw new Error('Framework not found');
 }
 
-init();
+const handler = registerRoutes({ framework });
 
 export { handler as GET, handler as POST, handler as PUT };
