+++ packages/admin/src/app/records/[integration]/page.tsx
@@ -1,10 +1,9 @@
 import { redirect } from 'next/navigation';
 
-import { getFramework } from '@/lib/framework-utils';
+import { framework } from '@/lib/framework-utils';
 
 export default async function Integration({ params }: { params: { integration: string } }) {
   const integrationName = params.integration.toUpperCase();
-  const framework = await getFramework();
   const integration = framework?.getIntegration(integrationName);
 
   if (!integration) {
