+++ packages/admin/src/domains/playground/server-actions/execute-framework-actions.ts
@@ -2,7 +2,7 @@
 
 import type { IntegrationActionExcutorParams } from '@arkw/core';
 
-import { getFramework } from '@/lib/framework-utils';
+import { framework } from '@/lib/framework-utils';
 
 interface Props {
   integrationName?: string;
@@ -11,8 +11,6 @@ interface Props {
 }
 
 export async function executeFrameworkAction(props: Props): Promise<void> {
-  const framework = await getFramework();
-
   if (!framework) {
     throw new Error('Framework not found');
   }
