+++ packages/admin/src/lib/framework-utils.ts
@@ -2,28 +2,16 @@ import { createFramework } from '@arkw/core';
 import fs from 'fs';
 import path from 'path';
 
-export async function getFramework() {
+function getFramework() {
   try {
-    const configFilePath = getConfigPath();
-    console.log('configFilePath:', configFilePath);
-    //const { config } = await import('/Users/shanethomas/dev/kepler/future/packages/admin/example.future.config');
-    const { config } = require(configFilePath);
-
+    const { config } = require(process.env.CONFIG_PATH!);
     return createFramework(config);
   } catch (error) {
     console.error('Error loading config:', error);
   }
 }
 
-export function getConfigPath() {
-  if (process.env.ARK_APP_DIR) {
-    const configPath = path.resolve(process.env.ARK_APP_DIR, 'arkw.config');
-    if (fs.existsSync(configPath)) {
-      return configPath;
-    }
-  }
-  return path.resolve(process.cwd(), 'example.future.config');
-}
+export const framework = getFramework();
 
 // @todo: the .env file should be able to be set to .env.local somehow
 // possibly defined in the config file?
