+++ packages/admin/src/lib/hooks/use-background-canvas-scroll.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { MouseEventHandler, RefObject, useCallback, useEffect, useState } from 'react';
 
 interface useBackgroundCanvasScrollProps {
