+++ packages/admin/src/lib/hooks/use-update-meta-tags.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { useEffect } from 'react';
 
 export const useUpdateMetaTags = ({ title, description }: { title: string; description?: string }) => {
